+++ packages/ecmascript-runtime-server/runtime.js
@@ -21,5 +21,6 @@
 // Note that the es6.reflect.* and es6.typed.* modules have been commented
 // out for bundle size reasons.
 
-require("core-js/modules/es7.string.pad-start");
-require("core-js/modules/es7.string.pad-end");
+require("core-js/modules/es.string.pad-start");
+require("core-js/modules/es.string.pad-end");
+require("core-js/modules/es.symbol.async-iterator");